Refactory: Do some abstraction of Service
1. previously all resource has its own Service layer.
2. after refactorying, they share the same Service layer.
